Be coherent: use long instead of int in the indirect access size variable

+++ src/RunTime/include/RunTime/local_data.hpp
@@ -80,7 +80,7 @@ protected:
 	std::set<int> indirect_access;
 
 	/// Number of indirect accesses
-	int indirect_access_size;
+	long indirect_access_size;
 
 	/// Same as direct_access_array but for indirect accesses
 	int* indirect_access_array;
